% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLm.R
\name{myLm}
\alias{myLm}
\title{myLm}
\usage{
myLm(formula, data)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic expression of the model being fitted.}

\item{data}{data frame containing variables in the modal being fitted.}
}
\value{
a list consisting of Coefficients and their corresponding standard error, t statistic and p-value. R squared and f statistic.
}
\description{
mimic the lm() function in R, used to fit linear modals. output according statistics given the data and formula being provided
}
\examples{
fit <- myLm(formula = Murder ~ Rape, data = USArrests)

fit$Coefficients
fit$Fstatistic

}
